Delete more code
