--- src/kernels/jupyter/session/backingFileCreator.web.ts
@@ -1,8 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-import { injectable } from 'inversify';
-import { BaseBackingFileCreator } from './backingFileCreator.base';
-
-@injectable()
-export class BackingFileCreator extends BaseBackingFileCreator {}
